% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Create a graph of a Country's Population}
\usage{
CountryPopulation(country = "United States of America")
}
\arguments{
\item{country}{String name of the desired country, defaults to USA}
}
\value{
The created ggplot of 'country' population data
}
\description{
Creates a ggplot graph object with a scatterplot of the specified
country's population from 1950 to 2020. Returns an error if the
country's name is not present in the WorldPopulation dataset.
}
\examples{
CountryPopulation()
plot <- CountryPopulation("Mexico")
}
