% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World Cup attendance dataset}
\format{
A data frame with 22 observations with 3 columns.
\describe{
\item{TotalAttend}{int: total number of people who attended the tournament}
\item{AvgAttend}{num: daily average number of people who attended the tournament}
\item{WorldCup}{chr: character string denoting which world cup this observation
pertains to. In format "\if{html}{\out{<Country>}}\if{html}{\out{<Year>}}"}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
A dataset of the average and total attendance of the world cup
tournaments from 1930 to 2022.
}
\keyword{datasets}
