% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population Dataset}
\format{
A data frame with 16685 observations with 3 columns.
\describe{
\item{Country}{chr: name of the country}
\item{Year}{int: Year of the observation from 1950 to 2020}
\item{Population}{int: recorded population for the country that year (in thousands)}
}
}
\usage{
WorldPopulation
}
\description{
A dataset of the population of individual countries from 1950 - 2020
Sorted by geographical location/region. Each year is an individual column,
and each row is a country
}
\keyword{datasets}
